refactor: Extract extension onboarding URL check to a separate function
